+++ app/controllers/application_controller.rb
@@ -1,2 +1,9 @@
 class ApplicationController < ActionController::API
+  def serialize_collection(collection, serializer, **options)
+    ActiveModel::Serializer::CollectionSerializer.new(collection, serializer:, **options).as_json
+  end
+
+  def render_errors(result)
+    render json: { errors: result[:errors] }, status: :unprocessable_entity
+  end
 end