+++ app/serializers/task_serializer.rb
@@ -1,0 +1,3 @@
+class TaskSerializer < ApplicationSerializer
+  attributes :id, :title, :description, :scheduled_at, :status, :created_at, :updated_at
+end