+++ app/services/tasks/show_service.rb
@@ -1,0 +1,11 @@
+module Tasks
+  class ShowService < ApplicationService
+    def initialize(id)
+      @id = id
+    end
+
+    def call
+      { success: true, task: Task.find(@id) }
+    end
+  end
+end